refactor: :art: organized files
