+++ src/components/PageContent/PageContent.astro
@@ -1,6 +1,6 @@
 ---
 import MoreMenu from '../RightSidebar/MoreMenu.astro';
-import TableOfContents from '../RightSidebar/TableOfContents.tsx';
+import TableOfContents from '../react/TableOfContents';
 
 const { content, githubEditUrl } = Astro.props;
 const title = content?.title;
