+++ src/components/RightSidebar/MoreMenu.astro
@@ -1,5 +1,5 @@
 ---
-import ThemeToggleButton from './ThemeToggleButton.tsx';
+import ThemeToggleButton from '../react/ThemeToggleButton';
 ---
 <div style="margin: 2rem 0; text-align: center;">
 	<ThemeToggleButton client:visible />
