+++ src/components/RightSidebar/RightSidebar.astro
@@ -1,5 +1,5 @@
 ---
-import TableOfContents from './TableOfContents.tsx';
+import TableOfContents from '../react/TableOfContents';
 import MoreMenu from './MoreMenu.astro';
 const { content } = Astro.props;
 const headers = content?.astro?.headers || [];
