+++ src/components/react/TableOfContents/hook.ts
@@ -1,0 +1,28 @@
+import { useState, useEffect, useRef } from 'react';
+
+
+export default function useTableOfContents(){
+  const itemOffsets = useRef([]);
+	const [activeId, setActiveId] = useState<string>();
+
+	useEffect(() => {
+		const getItemOffsets = () => {
+			const titles = document.querySelectorAll('article :is(h1, h2, h3, h4)');
+			itemOffsets.current = Array.from(titles).map((title) => ({
+				id: title.id,
+				topOffset: title.getBoundingClientRect().top + window.scrollY,
+			}));
+		};
+
+		getItemOffsets();
+		window.addEventListener('resize', getItemOffsets);
+
+		return () => {
+			window.removeEventListener('resize', getItemOffsets);
+		};
+	}, []);
+
+  return {
+    activeId
+  }
+}
