+++ src/components/react/ThemeToggleButton/index.tsx
@@ -1,5 +1,5 @@
-import { useState, useEffect } from 'react';
-import './ThemeToggleButton.css';
+import useTheme from "./hook"
+import './style.css';
 
 const themes = ['light', 'dark'];
 
@@ -17,27 +17,8 @@ const icons = [
 ];
 
 const ThemeToggle = () => {
-	const [theme, setTheme] = useState(() => {
-		if (import.meta.env.SSR) {
-			return undefined;
-		}
-		if (typeof localStorage !== 'undefined' && localStorage.getItem('theme')) {
-			return localStorage.getItem('theme');
-		}
-		if (window.matchMedia('(prefers-color-scheme: dark)').matches) {
-			return 'dark';
-		}
-		return 'light';
-	});
 
-	useEffect(() => {
-		const root = document.documentElement;
-		if (theme === 'light') {
-			root.classList.remove('theme-dark');
-		} else {
-			root.classList.add('theme-dark');
-		}
-	}, [theme]);
+  const {theme, setTheme} = useTheme()
 
 	return (
 		<div className="theme-toggle">
